@@ -1,15 +1,14 @@
-﻿namespace Asos.DotNetCore.Auth.Api.Demo.Models
+﻿namespace Asos.DotNetCore.Auth.Api.Demo.Models;
+
+public class ClaimDetailResponse
 {
-    public class ClaimDetailResponse
+    public ClaimDetailResponse(string claimType, string claimValue)
     {
-        public ClaimDetailResponse(string claimType, string claimValue)
-        {
-            ClaimType = claimType;
-            ClaimValue = claimValue;
-        }
-
-        public string ClaimType { get; set; }
-
-        public string ClaimValue { get; set; }
+        ClaimType = claimType;
+        ClaimValue = claimValue;
     }
+
+    public string ClaimType { get; }
+
+    public string ClaimValue { get; }
 }
