@@ -1,18 +1,17 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
-namespace Asos.DotNetCore.Auth.Api.Demo
-{
-   public class Program
-   {
-      public static void Main(string[] args)
-      {
-         CreateWebHostBuilder(args).Build().Run();
-      }
+namespace Asos.DotNetCore.Auth.Api.Demo;
 
-      public static IWebHostBuilder CreateWebHostBuilder(string[] args)
-      {
-         return WebHost.CreateDefaultBuilder(args).UseStartup<Startup>();
-      }
-   }
+public class Program
+{
+    public static void Main(string[] args)
+    {
+        CreateWebHostBuilder(args).Build().Run();
+    }
+
+    public static IWebHostBuilder CreateWebHostBuilder(string[] args)
+    {
+        return WebHost.CreateDefaultBuilder(args).UseStartup<Startup>();
+    }
 }
