@@ -1,17 +1,16 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 
-namespace Asos.DotNetCore.Auth.Api.Demo.Security
+namespace Asos.DotNetCore.Auth.Api.Demo.Security;
+
+public class SubjectMustMatchRouteParameterRequirement : IAuthorizationRequirement
 {
-    public class SubjectMustMatchRouteParameterRequirement : IAuthorizationRequirement
+    public SubjectMustMatchRouteParameterRequirement(string identifierClaim, string routeName)
     {
-        public SubjectMustMatchRouteParameterRequirement(string identifierClaim, string routeName)
-        {
-            IdentifierClaim = identifierClaim;
-            RouteName = routeName;
-        }
-
-        public string IdentifierClaim { get; set; }
-
-        public string RouteName { get; set; }
+        IdentifierClaim = identifierClaim;
+        RouteName = routeName;
     }
+
+    public string IdentifierClaim { get; }
+
+    public string RouteName { get; }
 }
